added donut chart
moved helper functions to a new file and added donut chart for inventory
