+++ dvpapp/src/components/Inventory/centered-container.css
@@ -1,0 +1,14 @@
+/* centered-container.svelte */
+.centered-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    min-height: 100vh;
+  }
+  
+  .centered-container canvas,
+  .centered-container input {
+    margin: 10px;
+  }
+
